+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the convention described at
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+### Added
+- Add `fold()` function on `Result` that combines `map()` and `mapErr()`.
+
 ## [3.0.1] - 2020-03-18
 ### Changed
 - Fix the package description and code formatting.